% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_points.R
\name{arc_points}
\alias{arc_points}
\title{Create coordinates for an arc}
\usage{
arc_points(arc_length = 360, from = 0, units = "degrees", n_points = 180)
}
\arguments{
\item{arc_length}{length of the arc in the units specified}

\item{from}{position the arc starts from}

\item{units}{degrees (default) or radians, the units of the arc}

\item{n_points}{number of points in the  arc}
}
\value{
a data frame of coordinates for the arc (x, Y)
}
\description{
Create coordinates for an arc
}
\examples{
arc_points() # returns a circle of 180 points, clockwise from north scaled -1 < x < 1, -1 < y <1
# arc of 10 points, clockwise from east to north
plot( arc_points( arc_length = 270, from =90, n_points = 10 ), type='l')
}
