% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_shape.R
\name{arc_shape}
\alias{arc_shape}
\title{Create coordinates for an arc/donut shape with outer and inner radii}
\usage{
arc_shape(
  arc_length = 360,
  from = 0,
  units = "degrees",
  n_points = 180,
  radius_outer = 1,
  radius_inner = 0.5
)
}
\arguments{
\item{arc_length}{length of the arc in the units specified}

\item{from}{position the arc starts from}

\item{units}{degrees (default) or radians, the units of the arc}

\item{n_points}{number of points in the  arc}

\item{radius_outer}{radius of outer arc, should be > inner}

\item{radius_inner}{inner arc radius, should be < outer}
}
\value{
data frame of shape coordinates
}
\description{
Create coordinates for an arc/donut shape with outer and inner radii
}
