% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_labels.R
\name{axis_labels}
\alias{axis_labels}
\title{Draw axis labels (without axis line or ticks)}
\usage{
axis_labels(side, col = "grey20", cex = 1, ...)
}
\arguments{
\item{side}{integer specifying which side of the plot the axis is drawn on: 1=below, 2=left, 3=above and 4=right}

\item{col}{colours of the labels}

\item{cex}{size of labels following the cex, character expansion argument of par()}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw axis labels (without axis line or ticks)
}
