% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_line.R
\name{axis_line}
\alias{axis_line}
\title{Draw line for axis (without labels or axis ticks)}
\usage{
axis_line(side = NULL, col = "grey20", ...)
}
\arguments{
\item{side}{integer specifying which side of the plot the axis is drawn on: 1=below, 2=left, 3=above and 4=right}

\item{col}{colour of the lines}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw line for axis (without labels or axis ticks)
}
