% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_ticks.R
\name{axis_ticks}
\alias{axis_ticks}
\title{Draw axis tick marks (without labels or axis line)
#'}
\usage{
axis_ticks(side = NULL, lwd = 1, col = "grey20", length = -0.05, ...)
}
\arguments{
\item{side}{integer specifying which side of the plot the axis is drawn on: 1=below, 2=left, 3=above and 4=right}

\item{lwd}{number for the line width}

\item{col}{colour of the tick marks}

\item{length}{distance of tick marks into the plot region (+ve values) or out of the plot region (-ve values)}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw axis tick marks (without labels or axis line)
#'
}
