% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines-package.R
\docType{package}
\name{baselines-package}
\alias{baselines}
\alias{baselines-package}
\title{baselines: Calculates Distance From Desired Baseline}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Builds off the vegan package with functions we have been using with paleo data. Calculated the ellipse around a set of points and calculates minimum distance from points to nearest edge of ellipse.
}
\author{
\strong{Maintainer}: Olivia Burge \email{olivia.burge@gmail.com}

}
\keyword{internal}
