% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_outer.R
\name{box_outer}
\alias{box_outer}
\title{Draw box around figure region}
\usage{
box_outer(border = "hotpink", lwd = 2, lty = "33", ...)
}
\arguments{
\item{border}{colour of border line}

\item{lwd}{line width of border line}

\item{lty}{line type of border line}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw box around figure region
}
