% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistanceStart.R
\name{calculateDistanceStart}
\alias{calculateDistanceStart}
\title{Caculate ecological distance from first sample}
\usage{
calculateDistanceStart(
  speciesData,
  metaData,
  idCol,
  idColType,
  distMethod,
  threshold = 0
)
}
\arguments{
\item{speciesData}{dataset with species abundance, pre-transformed if desired}

\item{metaData}{dataset with metadata, such as depth and age}

\item{idColType}{need to specify whether the idcol is a depth style or year style variable.}

\item{distMethod}{the distance method to use. See ?vegan::vegdist for options.}

\item{threshold}{keeps species only where those with a total summed abundance (not frequency) of more than or equal to the threshold. Default is zero.}
}
\value{
A dataframe with idCol and the distance for each point, including the distance from the first sample to itself.
}
\description{
Caculate ecological distance from first sample
}
\examples{

## This example uses data from the analogue package

require(analogue)
data(abernethy)
speciesData <- abernethy[!names(abernethy) \%in\% c("Depth", "Age")]
metaData <- abernethy[c("Depth", "Age")]

distFirst <- calculateDistanceStart(speciesData, metaData,
idCol = "Age", distMethod = "jaccard")

with(distFirst, plot(x = Age, y = dist_jaccard,
ylab = "Jaccard distance"))

}
