% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_arrow.R
\name{canvas_arrow}
\alias{canvas_arrow}
\title{Draw arrow over the canvas (device) region as coordiantes in ranges 0 < x < 1, 0 < y < 1}
\usage{
canvas_arrow(
  x1 = 0.1,
  y1 = 0.1,
  x2 = 0.7,
  y2 = 0.3,
  col = "lemonchiffon3",
  lwd = 2,
  ...
)
}
\arguments{
\item{x1}{coordinate of x start point}

\item{y1}{coordinate of y start point}

\item{x2}{coordinate of x end point}

\item{y2}{coordinate of y end point}

\item{col}{colour of arrow}

\item{lwd}{arrow line widths}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw arrow over the canvas (device) region as coordiantes in ranges 0 < x < 1, 0 < y < 1
}
