% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_background.R
\name{canvas_background}
\alias{canvas_background}
\title{Add a background colour to the canvas, whole device}
\usage{
canvas_background(col = adjustcolor("orchid", alpha.f = 0.4), ...)
}
\arguments{
\item{col}{the colour for both the background and border}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Add a background colour to the canvas, whole device
}
