% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_grid.R
\name{canvas_grid}
\alias{canvas_grid}
\title{Draw a grid overlaying the canvas to determine object poistions and coordinates}
\usage{
canvas_grid(x1 = 0, x2 = 1, y1 = 0, y2 = 1, col = "#dbe4f0", lwd = 0.5, ...)
}
\arguments{
\item{x1}{left position of the grid (defaulting to whole device)}

\item{x2}{right grid position (defaulting to whole device)}

\item{y1}{bottom position of grid (defaulting to whole device)}

\item{y2}{top position of grid (defaulting to whole device)}

\item{col}{colour for the grid with 0.5 transparency applied}

\item{lwd}{line width for grid}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw a grid overlaying the canvas to determine object poistions and coordinates
}
\examples{
plot(1:9)
canvas_grid()
}
