% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_rect.R
\name{canvas_rect}
\alias{canvas_rect}
\title{Draw a rectangle to the canvas/device region (rather than the plot region)}
\usage{
canvas_rect(
  x1 = 0.1,
  y1 = 0.5,
  x2 = 0.9,
  y2 = 0.9,
  border = T,
  col = "lemonchiffon",
  angle = NULL,
  ...
)
}
\arguments{
\item{x1}{left coordinate of rectangle}

\item{y1}{bottom coordinate}

\item{x2}{right coordinate}

\item{y2}{top coordinate}

\item{border}{colour of the border line}

\item{col}{fill colour of rect}

\item{angle}{angle (in degrees) of the shading lines.}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw a rectangle to the canvas/device region (rather than the plot region)
}
