% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_shape_above.R
\name{close_shape_above}
\alias{close_shape_above}
\title{Create a shape for a polygon upwards from your data}
\usage{
close_shape_above(shape, left_to_right = T, extreme_value = NULL)
}
\arguments{
\item{shape}{values for the coordinates of the data/shape to be closed upwards}

\item{left_to_right}{a logical parameter specifying whether the data is read left to right or not}

\item{extreme_value}{a value defining the top of the new shape}
}
\value{
a data frame of coordinates for the new, closed shape
}
\description{
Create a shape for a polygon upwards from your data
}
