% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_shape_right.R
\name{close_shape_right}
\alias{close_shape_right}
\title{Create a shape for a polygon to the right of your data}
\usage{
close_shape_right(shape, top_to_bottom = T, extreme_value = NULL)
}
\arguments{
\item{shape}{values for the coordinates of the data/shape to be closed to the right}

\item{extreme_value}{a value defining the top of the new shape}

\item{left_to_right}{a logical parameter specifying whether the data is read left to right or not}
}
\value{
a data frame of coordinates for the new, closed shape
}
\description{
Create a shape for a polygon to the right of your data
}
