% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_2_hex.R
\name{col_2_hex}
\alias{col_2_hex}
\title{Get hex code for colour/hex name}
\usage{
col_2_hex(col)
}
\arguments{
\item{col}{colour to convert to hex code}
}
\description{
Get hex code for colour/hex name
}
\examples{
col_2_hex("hotpink")
}
