% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_2_hsv.R
\name{col_2_hsv}
\alias{col_2_hsv}
\title{Get hsv values for colour/hex name}
\usage{
col_2_hsv(col)
}
\arguments{
\item{col}{colour to convert to hsv}
}
\description{
Get hsv values for colour/hex name
}
\examples{
col_2_hsv("slateblue2")
col_2_hsv("slateblue2")[1]
}
