% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_cutter.R
\name{colour_cutter}
\alias{colour_cutter}
\title{Create colour bins using quantile, range, sd and Kmeans methods}
\usage{
colour_cutter(
  data = 1:7,
  colours = grey.colors(7),
  cutter = "quantile",
  n_nested_bins = 2
)
}
\arguments{
\item{data}{the data values to be binned}

\item{colours}{list of colours}

\item{cutter}{the method to be used}

\item{n_nested_bins}{number of bins within bins for visual testing}
}
\value{
an S3 auto-shading object emulating that of the GISTOOLS package
}
\description{
Create colour bins using quantile, range, sd and Kmeans methods
}
