% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palette.R
\name{colour_palette}
\alias{colour_palette}
\title{Display a list of colours with hex codes and against backgrounds}
\usage{
colour_palette(colours = grey.colors(15))
}
\arguments{
\item{colours}{list of colours to be displayed}
}
\description{
Display a list of colours with hex codes and against backgrounds
}
\examples{
colour_palette()
}
