% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees_2_radians.R
\name{degrees_2_radians}
\alias{degrees_2_radians}
\title{Convert degrees to radians}
\usage{
degrees_2_radians(degrees)
}
\arguments{
\item{degrees}{the values to be converted}
}
\description{
Convert degrees to radians
}
