% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_centre.R
\name{get_plot_centre}
\alias{get_plot_centre}
\title{Retrieve centre point of plot}
\usage{
get_plot_centre(x_prop = 0.5, y_prop = 0.5)
}
\arguments{
\item{x_prop}{the x position relative to x1 (0) and x2 (1)}

\item{y_prop}{the x position relative to x1 (0) and x2 (1)}
}
\description{
Retrieve centre point of plot
}
