% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_device_quartz.R,
%   R/graphics_device_windows.R
\name{graphics_device_quartz}
\alias{graphics_device_quartz}
\title{Set quartz device (mac) as default graphics device in options(...)}
\usage{
graphics_device_quartz()

graphics_device_quartz()
}
\description{
Set quartz device (mac) as default graphics device in options(...)

Set quartz device (mac) as default graphics device in options(...)
}
