% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_device_rstudio.R,
%   R/graphics_device_windows.R
\name{graphics_device_rstudio}
\alias{graphics_device_rstudio}
\title{Set RStudioGD device as default graphics device in options(...)}
\usage{
graphics_device_rstudio()

graphics_device_rstudio()
}
\description{
Set RStudioGD device as default graphics device in options(...)

Set RStudioGD device as default graphics device in options(...)
}
