% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_device_windows.R
\name{graphics_device_windows}
\alias{graphics_device_windows}
\title{Set windows device as default graphics device in options(...)}
\usage{
graphics_device_windows()
}
\description{
Set windows device as default graphics device in options(...)
}
