% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_device_windows.R,
%   R/graphics_device_x11.R
\name{graphics_device_x11}
\alias{graphics_device_x11}
\title{Set x11 device as default graphics device in options(...)}
\usage{
graphics_device_x11()

graphics_device_x11()
}
\description{
Set x11 device as default graphics device in options(...)

Set x11 device as default graphics device in options(...)
}
