% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inches_2_cm.R
\name{inches_2_cm}
\alias{inches_2_cm}
\title{Convert inches to cm}
\usage{
inches_2_cm(inches)
}
\arguments{
\item{cms}{values in inches to be converted to centimetres}
}
\description{
Convert inches to cm
}
