% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_as_segments.R
\name{lines_as_segments}
\alias{lines_as_segments}
\title{Draw lines piecewise, to add widths or colours for each segment}
\usage{
lines_as_segments(x, y, ...)
}
\arguments{
\item{x}{values for the x coordinates}

\item{y}{values for the y coordinates}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw lines piecewise, to add widths or colours for each segment
}
