% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_from_shapes.R
\name{mask_from_shapes}
\alias{mask_from_shapes}
\title{Cut one shape form another to create a mask}
\usage{
mask_from_shapes(
  inner_shape,
  outer_shape = plot_region_rect(),
  draw = T,
  border = F,
  col = "skyblue4",
  ...
)
}
\arguments{
\item{inner_shape}{coordinates of shape to be cut out}

\item{outer_shape}{coordinates of outer shape which is cut from, defaults to plot region}

\item{draw}{logical parameter defining whether shape is drawn or not, defaults to TRUE}

\item{border}{border colour if shape is drawn}

\item{col}{fill colour if shape is drawn}

\item{...}{}
}
\value{
coordinates of mask shape and shape drawn if draw=T (default)
}
\description{
Cut one shape form another to create a mask
}
