% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_outer_margin.R
\name{overlay_outer_margin}
\alias{overlay_outer_margin}
\title{Overlay outer margins on an active plot device, visual query}
\usage{
overlay_outer_margin()
}
\value{
overlay the outer margin region and margin lines on active plot
}
\description{
Overlay outer margins on an active plot device, visual query
}
\examples{
plot_test()
overlay_outer_margin()
}
