% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_measures.R
\name{par_measures}
\alias{par_measures}
\title{Create conversion measures ( cm, inch, px ) for margins, and plot, figure and device regions}
\usage{
par_measures(...)
}
\arguments{
\item{...}{global par() parameters passed to function}
}
\value{
s3 object of par measures with par values appended
}
\description{
Create conversion measures ( cm, inch, px ) for margins, and plot, figure and device regions
}
