% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyllotaxis_points.R
\name{phyllotaxis_points}
\alias{phyllotaxis_points}
\title{Generate points in a phyllotaxis pattern for visual demonstrations}
\usage{
phyllotaxis_points(n_points = 1000, ma)
}
\arguments{
\item{n_points}{number of points to create}

\item{ma}{value controlling the spread of the pattern}
}
\value{
a data frame of values for each point
}
\description{
Generate points in a phyllotaxis pattern for visual demonstrations
}
