% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_background.R
\name{plot_background}
\alias{plot_background}
\title{Add a background colour to the plot region}
\usage{
plot_background(col = adjustcolor("skyblue2", alpha.f = 0.5), ...)
}
\arguments{
\item{col}{the colour for the background and border}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Add a background colour to the plot region
}
