% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blank.R
\name{plot_blank}
\alias{plot_blank}
\title{Create a blank plot}
\usage{
plot_blank(x = 0:1, y = 0:1, margins = T, ...)
}
\arguments{
\item{x}{values for x, as a sequence or vector, max & min values are used for data range}

\item{y}{values for y, as a sequence or vector, max & min values are used for data range}

\item{margins}{logical value specifying if margins are preserved (default), or not}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Create a blank plot
}
