% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test.R
\name{plot_test}
\alias{plot_test}
\title{Create a test plot ( with data and inner and outer margins)}
\usage{
plot_test()
}
\description{
Create a test plot ( with data and inner and outer margins)
}
\examples{
plot_test()
box_outer()
overlay_outer_margin()
axis_labels( side=4, col="hotpink", cex=2 )
overlay_margin()
}
