% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_between_points.R
\name{points_between_points}
\alias{points_between_points}
\title{Create N divisions between two points (straight line)}
\usage{
points_between_points(point_1, point_2, distance = 0.5)
}
\arguments{
\item{point_1}{coordinates of first point c( x1 ,y1 )}

\item{point_2}{coordinates of second point c( x2, y2 )}

\item{distance}{value(s) for distances (proportion) of positions between points}
}
\value{
data frame of coordinates for the points, between points
}
\description{
Create N divisions between two points (straight line)
}
