% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians_2_degrees.R
\name{radians_2_degrees}
\alias{radians_2_degrees}
\title{Convert radians to degrees}
\usage{
radians_2_degrees(radians)
}
\arguments{
\item{radians}{the values to be converted}
}
\description{
Convert radians to degrees
}
