% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect_fade.R
\name{rect_fade}
\alias{rect_fade}
\title{Draw a rect with faded fill between two or more colours}
\usage{
rect_fade(
  x1 = par()$usr[1],
  y1 = par()$usr[3],
  x2 = par()$usr[2],
  y2 = par()$usr[4],
  start_side = 1,
  resolution = 20,
  colours = c("grey20", "hotpink", "grey80"),
  border_colour = F,
  xpd = F,
  alpha = NULL,
  ...
)
}
\arguments{
\item{x1}{value for left, horizontal position in plot region coordinates}

\item{y1}{value for bottom, vertical position in plot region coordinates}

\item{x2}{value for right}

\item{y2}{value for top}

\item{start_side}{number specifying which axis side the fade begins from}

\item{resolution}{number of bins in the fade}

\item{colours}{two or more colours for the fade}

\item{border_colour}{colour of the border line}

\item{xpd}{logical value defining if rect is drawn outside plot region}

\item{alpha}{transaprency value}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw a rect with faded fill between two or more colours
}
