% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect_rounded.R
\name{rect_rounded}
\alias{rect_rounded}
\title{create a rect with rounded corners}
\usage{
rect_rounded(x1 = -1, y1 = -1, x2 = 1, y2 = 1, rounding = 0.5, n_points = 180)
}
\arguments{
\item{x1}{left x coordinate (defaults to -1 to match arc_points data range)}

\item{y1}{bottom y coordinate (defaults to -1 to match arc_points data range)}

\item{x2}{right x coordinate (defaults to 1 to match arc_points data range)}

\item{y2}{top y coordinate (defaults to 1 to match arc_points data range)}

\item{rounding}{similarity to circle, 1 = circle & 0 = rect}

\item{n_points}{number of points in circle, or ceiling( n_points/4 ) for rounded corner}
}
\description{
create a rect with rounded corners
}
