% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect_sides.R
\name{rect_sides}
\alias{rect_sides}
\title{Draw square brackets, lines, box and other side combinations around rect}
\usage{
rect_sides(
  x1,
  y1,
  x2,
  y2,
  sides = c(1, 2, 3, 4),
  col = NULL,
  border = "black",
  ...
)
}
\arguments{
\item{x1}{value for left, horizontal position in plot region coordinates}

\item{y1}{value for bottom, vertical position in plot region coordinates}

\item{x2}{value for right}

\item{y2}{value for top}

\item{sides}{the sides to be drawn in the form c( 1, 2, 3, 4 ), 1:3 etc.}

\item{col}{fill colour of the rect}

\item{border}{border line colour}

\item{...}{global par() parameters}
}
\value{
draws a rect with teh specified sides as lines
}
\description{
Draw square brackets, lines, box and other side combinations around rect
}
