% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_figure_region.R
\name{restore_figure_region}
\alias{restore_figure_region}
\title{Restore the figure region parameters (but not margins)}
\usage{
restore_figure_region(previous_par)
}
\arguments{
\item{previous_par}{}
}
\description{
Restore the figure region parameters (but not margins)
}
