% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_par.R
\name{restore_par}
\alias{restore_par}
\title{Restore par settings to defaults or to specified par values}
\usage{
restore_par(previous_par = NULL, ...)
}
\arguments{
\item{previous_par}{an object contained saved par values}

\item{...}{other parameters from par() parameter}
}
\description{
Restore par settings to defaults or to specified par values
}
