% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotonuiaha}
\alias{rotonuiaha}
\title{Rotonuiaha pollen counts}
\format{A data frame with 73 rows and 93 variables:
\describe{
  \item{site}{site name}
  \item{year}{year, derived from radiocarbon dating}
  ...
}}
\usage{
rotonuiaha
}
\description{
A dataset containing pollen counts for Rotonuiaha
}
\keyword{datasets}
