% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_direction.R
\name{shape_direction}
\alias{shape_direction}
\title{Find direction of rotation for shape coordinates}
\usage{
shape_direction(shape, draw = FALSE)
}
\arguments{
\item{shape}{coordinates for shape}

\item{draw}{logical value specifying whether shape is drawn or not, defaults to false}
}
\description{
Find direction of rotation for shape coordinates
}
