% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_box_plot.R
\name{text_box_plot}
\alias{text_box_plot}
\title{Draw text box in plot region}
\usage{
text_box_plot(
  x1 = 0.3,
  y1 = 0.3,
  x2 = 0.7,
  y2 = 0.7,
  text = NULL,
  cex = 1,
  font = 1,
  family = "serif",
  adj = 0,
  srt = 0,
  border = F,
  lty = 1,
  xpd = T,
  col_box = "grey97",
  col_text = "grey30",
  left_margin = 0.075,
  right_margin = 0.075,
  top_margin = 0.025,
  bottom_margin = 0.05,
  line_spacing = 0.05,
  vertical_alignment = "top",
  horizontal_alignment = "left",
  ...
)
}
\arguments{
\item{x1}{value for left, horizontal position in plot region coordinates}

\item{y1}{value for bottom, vertical position in plot region coordinates}

\item{x2}{value for right}

\item{y2}{value for top}

\item{text}{a character vector or expression specifying the text to be written with "Lorem ipsum..." default}

\item{cex}{text size, character expansion value (see par)}

\item{font}{default to the values of the global graphical parameters in par()}

\item{family}{default to the values of the global graphical parameters in par()}

\item{adj}{alignment of text (but see arguments below)}

\item{srt}{rotation (not currently implemented)}

\item{border}{logical or value for the colour of the box border}

\item{lty}{value for the line type, e.g. 1, 2, 3 or 4... or "22"}

\item{col_box}{fill colour of the box}

\item{col_text}{colour of the text}

\item{left_margin}{padding around text within box}

\item{right_margin}{padding around text within box}

\item{top_margin}{padding around text within box}

\item{bottom_margin}{padding around text within box}

\item{line_spacing}{padding between lines}

\item{vertical_alignment}{top, centre or bottom alignment}

\item{horizontal_alignment}{left, middle or right alignment}

\item{...}{other parameters, see global graphical parameters in par()}
}
\description{
Draw text box in plot region
}
