% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_main_title.R
\name{text_main_title}
\alias{text_main_title}
\title{Add a main title to your plot}
\usage{
text_main_title(text = "title", col = "grey30", cex = 1.2, font = 2, ...)
}
\arguments{
\item{text}{the text to be displayed}

\item{col}{the colour of that text}

\item{cex}{the size of the text (character explansion)}

\item{font}{the text formatting - 1, plain; 2, bold; 3, italics; 4, bold italics.}

\item{...}{global par() parameters}
}
\description{
Add a main title to your plot
}
\examples{
text_main_title()
}
