% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_sub_title.R
\name{text_sub_title}
\alias{text_sub_title}
\title{Add a sub title to your plot}
\usage{
text_sub_title(text = "sub title", col = "grey30", cex = 1, font = 1, ...)
}
\arguments{
\item{text}{the text to be displayed}

\item{col}{the colour of that text}

\item{cex}{the size of the text (character explansion)}

\item{font}{the text formatting - 1, plain; 2, bold; 3, italics; 4, bold italics.}

\item{...}{global par() parameters}
}
\description{
Add a sub title to your plot
}
\examples{
text_sub_title()
}
