% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_x_label.R
\name{text_x_label}
\alias{text_x_label}
\title{Add an x axis label (side = 1)}
\usage{
text_x_label(text = "x label", col = "grey30", cex = 1, font = 1, ...)
}
\arguments{
\item{text}{the text to be displayed}

\item{col}{the colour of that text}

\item{cex}{the size of the text (character explansion)}

\item{font}{the text formatting - 1, plain; 2, bold; 3, italics; 4, bold italics.}

\item{...}{global par() parameters}
}
\description{
Add an x axis label (side = 1)
}
\examples{
text_x_label()
}
