% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_y_label.R
\name{text_y_label}
\alias{text_y_label}
\title{Add an y axis label (side = 2)}
\usage{
text_y_label(text = "y label", col = "grey30", cex = 1, font = 1, ...)
}
\arguments{
\item{text}{the text to be displayed}

\item{col}{the colour of that text}

\item{cex}{the size of the text (character explansion)}

\item{font}{the text formatting - 1, plain; 2, bold; 3, italics; 4, bold italics.}

\item{...}{global par() parameters}
}
\description{
Add an y axis label (side = 2)
}
\examples{
text_y_label()
}
