% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistanceStart.R
\name{calculateDistanceStart}
\alias{calculateDistanceStart}
\title{Caculate ecological distance from first sample}
\usage{
calculateDistanceStart(speciesData, metaData, idCol, distMethod)
}
\arguments{
\item{speciesData}{dataset with species abundance, pre-transformed if desired}

\item{metaData}{dataset with metadata, such as depth and age}

\item{idCol}{name of column in metaData by which speciesData is ordered. Generally depth or year. Needs to be in ascending order (ie getting deeper or older), in the same order as the species data.}

\item{distMethod}{the distance method to use. See ?vegan::vegdist for options.}
}
\value{
A dataframe with idCol and the distance for each point, including the distance from the first sample to itself.
}
\description{
Caculate ecological distance from first sample
}
\examples{

## This example uses data from the analogue package

require(analogue)
data(abernethy)
speciesData <- abernethy[!names(abernethy) \%in\% c("Depth", "Age")]
metaData <- abernethy[c("Depth", "Age")]

distFirst <- calculateDistanceStart(speciesData, metaData,
idCol = "Age", distMethod = "jaccard")

with(distFirst, plot(x = Age, y = dist_jaccard,
ylab = "Jaccard distance"))

}
