% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_blank.R
\name{canvas_blank}
\alias{canvas_blank}
\title{create a blank canvas to draw onto, with no margins (or outer margins) and whole device}
\usage{
canvas_blank(asp = NULL, restore_previous_par = F, ...)
}
\arguments{
\item{asp}{aspect ratio in the  canvas}

\item{restore_previous_par}{logical value specifying if previous par parametrs should be restored, defaulting to FALSE}

\item{...}{}
}
\description{
create a blank canvas to draw onto, with no margins (or outer margins) and whole device
}
