% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotonuiaha}
\alias{rotonuiaha}
\title{Rotonuiaha pollen counts}
\format{
A data frame with 73 rows and 96s variables. The site column just records this is from Rotonuiaha. The year column is derived from radiocarbon dating. The depth column is in cm. For details of the ecology of the core and site see Wilmshurst & McGlone (1996).
}
\source{
\url{http://onlinelibrary.wiley.com/doi/10.1111/cobi.12150/abstract}
}
\usage{
data(rotonuiaha)
}
\description{
A dataset containing pollen counts for Rotonuiaha, in the North Island of New Zealand.
}
\references{
Wilmshurst JM, McGlone MS. 1996. Forest disturbance in the central North Island, New Zealand, following the 1850 BP Taupo eruption. The Holocene 6:399–411. Available from http://journals.sagepub.com/doi/abs/10.1177/095968369600600402.
}
\keyword{datasets}
